moved all styling to its own folder for better organization
