+++ src/NavBar.jsx
@@ -1,4 +1,4 @@
-import './NavBar.css'
+import './styling/NavBar.css'
 
 export default function NavBar() {
   return (
