+++ src/Projects.jsx
@@ -1,4 +1,4 @@
-import './Projects.css'
+import './styling/Projects.css'
 
 function Projects() {
   return <div>
