+++ src/styling/NavBar.css
